Create Prime number till given number.c
Prime number
A prime number is divisible only by 1 and itself. You are given a positive integer. Write an algorithm to find all the prime numbers from 2 to the given positive number.



Input Format:

The input consists of an integer num, representing the number written on the board.

Output Format:

Print space-separated integers representing the numbers required by the teacher.